--- src/main/java/com/nulp/mobilepartsshop/api/v1/auth/dto/request/RegistrationRequest.java
@@ -1,21 +1,0 @@
-package com.nulp.mobilepartsshop.api.v1.auth.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class RegistrationRequest {
-
-    private String firstname;
-
-    private String lastname;
-
-    private String username;
-
-    private String password;
-}
