+++ src/main/java/com/nulp/mobilepartsshop/api/v1/authentication/dto/response/AuthorizationResponse.java
@@ -1,0 +1,15 @@
+package com.nulp.mobilepartsshop.api.v1.authentication.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class AuthorizationResponse {
+
+    private String jwtToken;
+}
